Choose active SIM on MSIM devices when making emergency calls.
We previously always chose the default Phone object for emergency calls, which
forced us to try to dial with the SIM in slot1 always. When the only SIM
is in slot 2, this caused the call not to go through.

Fixed to choose the SIM that was given to us and if that fails we move
on to the first active SIM and if that fails, *then* we opt for the
default phone.

Bug: 18746266
Change-Id: Ia005c1b2e9d1a08e9ab119f799552c97316f0853