Add error handling for outgoing WFC calls.
+ If POWER_OFF or OUT_OF_SERVICE is returned, show more specific
error messages based on WFC config and current wifi connectivity.

Assumes that IN_SERVICE will return if a call can be made,
regardless of the transport technology being used.

Assumes that OUT_OF_SERVICE will be returned if a call can not
be made and the radio is powered on.

Assumes that POWER_OFF will be returned if a call can not be
made and the radio is powered off.

Bug: 19167571
Change-Id: I32790bd05bb33e4890ee797f9f7a5297f6f0228c