Check radio, not ServiceState, for if radio is on.
Cherrypicked from m-wireless-telephony-dev branch.

For E911 calls, pre-empt user WFC settings by always turning on the
radio so we can make a best-effort attempt to complete the call.

Formerly, we checked the ServiceState for STATE_POWER_OFF to indicate
the state of the radio. However, with WFC, the ServiceState could
return IN_SERVICE if connected to a Wifi network when the radio is
powered off because of airplane mode.

Instead, now we check for the radio's power state directly, and power
it on if necessary using the EmergencyCallHelper before attempting
the outgoing call.

Bug: 19151398
Change-Id: I15bac137b402455177ad94e43da7aa18fa3f78c0