Save status and set up sync adapter.
There are two parts to this CL:
1. Set up the sync adapter - this involves relevant xml files and
additions to the manifest and subclassing the general VvmSyncAdapter.
2. Processing the status sms that activates the voicemail source -
saving it to the voicemail status database and registering a sync
account.
These were done together because in order to have a fully-functioning
sync adapter, at least one sync account must be registered for it.

Also moved the code from VvmSyncService to OmtpVvmSyncService because
there's no need to have a super class yet. Note that most of the syncing
logic was removed for now so it  can be added incrementally.

Bug: 19236241
Change-Id: If2baf43de4bd70761feadd2d2bd5c4f58c25d19b