Fix MSIM for CDMA and GSM Call Options.
This scenario was just... pretty broken before. Update, to pass
the right subscription ids to the right menus, display the right
strings, and obtain the correct phone reference for the sub id.

In particular, moved some initialization code into
GsmUmtsCallOptions so that initialization of its layout can be
shared between it and CallFeatureSettings.

Bug: 19301753
Change-Id: I00954c14f386f3e60b3824138b651c28a40a5c06