Download all message metadata from server.
When a new SIM is inserted, the following should happen:
1. All local data is synced up to the server
2. All local data is wiped
3. All non-deleted voicemail metadata is downloaded from the server
(this means that there is enough information to create entries for the
voicemails but not that the content/audio files are downloaded).

Also added back LogUtils in all the e-mail library code (everything
under com.android.phone.common.mail).

All newly added files are direct copies except for "Flag" (including
refactorings of the files that use "Flag", converting enums into static
strings) and Utility (which selects only the necessary utility
functions). Note that this means that some of the files have dead code
in them--so later CLs will remove the excess weighte

Bug: 19236241

Change-Id: Ibc49c96667db8ec0de2ab52bdd82bfd0f6993135