Prevent merging conference calls hosted on peer device.
With IMS, the "multiparty" bit on an ImsCall is set to "true" when a call
is merged into a conference.  This not only occurs on the device hosting
the conference call, but also on the devices of the callers merged into
the conference.

This CL changes the TelephonyConnection to listen to multipart state
changes reported by the ImsPhoneConnection.  It uses this to notify the
ImsConferenceController when a conference has started.  Note: the handler
is re-used to ensure the conference creation happens on the correct thread.
PhoneBase#checkCorrectThread would error out during creation of the
conference otherwise.

Modified ImsConferenceController to check for conferences which are not
conference hosts and to exclude them from the list of conferenceables.
We also exclude connections marked as multiparty which are not a conference
host (it is possible we have not yet turned these into a conference yet).

Bug: 19478784
Change-Id: I15c39e9c922c8a71e30f210f212e0ec779f5f939