Use telecom.Connection state for WFC StatusHint.
Attempting another rewrite of this logic, since the existence of
the original connection seemed to be inconsistent.

Instead, leverage other mechanisms to identify whether the call is
"incoming" or "active. Originally, I wanted to just check for
getState() == STATE_RINGING, but this seemed to be skipped.
Using isValidRingingCall() instead worked correctly though.

This is more consistent and fixes some janky SIP/account status
label formatting which used to be happening.

I'm not sure if it's necessary, but it also seemed safer to update
the status hints after class-level status update as well.

Bug: 19995266
Bug: 19970744
Change-Id: I5885da7ca6cb93e163a558f42c594d29b191c03c