Network type will reflect user's setting
Phone.setPreferredNetworkType will filter the netowrk mode based on
the modem's capability, before sending it to the modem.
As a result Phone.getPreferredNetworkType would not reflect what we
set and the API has been deprecated.

The network mode UI will show the user's preference rather than
what the modem is using.

Change-Id: I0900222806b5248f923bccf06c04cd70faba5370