Port VM settings to its own activity.
I gave up on PreferenceFragments because it's too hard to migrate
the Activity Dialog logic at this point. Moving voicemail settings
code to its own activity, with an eye towards avoiding any breakages
in the attends related to it.

I tried to rearrange some methods for better organization while I
was at it... all this logic is still in much need of clarity, but
I figure this is no worse of a state than before, and at least now
it's separated from the rest of the call settings.

No (deliberate) changes to logic besides that.

Bug: 17019623
Change-Id: I20aa177ca699ad2c823e49f0fed0dcd89f75596d