Cleanup of Video Call pause functionality.
Whether the paused videoState is available is dependent on the carrier's
implementation of the VT spec. The original VT implementation assumed
that this was stored in a system property accessed via InCall; these CLs
move this to a Call/Connection capability which will ultimately support
multisim video capable devices.

- For MCC/MNC 310/410, disable pause of ims video calls.
- For MCC/MNC 311/480, allow pause of ims video calls.
- Default for others is allow pause.

- Modified TelecomAccountRegistry to store flag indicating which accounts
allow pausing of video.  If account doesn't support video, pausing is
inherently not allowed either.
- Modified PstPhoneCapabilities notifier to add listener which calls back
to TelecomAccountRegistry#AccountEntry to notify when the video capability
has changed, triggering the AccountEntry to refresh itself accordingly.


Bug: 16680364
Bug: 19820114
Change-Id: I3840f92270100811161120dffcfe297bef7c4ea2