Phone app might use cleartext network traffic.
This declares that the Phone app might be using cleartext network
traffic (e.g., HTTP instead of HTTPS). This CL does not change the
behavior of the app. The flag added by this CL to the app's manifest
will make it easier to identify apps that use cleartext traffic in
the future.

The only reason for using cleartext traffic is that the app can fetch
voicemails over cleartext IMAP instead of IMAP over TLS/SSL or IMAP
with STARTTLS.

Bug: 19215516
Change-Id: Ic6b439df5b89bcfbef6e982b1dfbac662daebae5