Make voicemail IMAP requests through data connection.
Most IMAP servers for visual voicemail can only be accessed through the
carrier's data network. This is the first part of the change that
ensures that voicemail fetches are done through the data connection.
This part of the change also updates the plumbing through the IMAP code
to get the correct socket corresponding to the network.
The change to ensure all syncing is done through the data connection is
coming soon.
Also trigger a sync every time a STATUS sms message is received because
it could mean the credentials were refreshed and therefore the local
data needs to be resynced.

Bug: 20345518
Change-Id: Ib6d9d508e38823849ee37459071d40da622ae720