Persist user enable/disable setting through reboot.
When the carrier vvm app is installed but the user manually enables
dialer vvm, we should ensure that the user preference is persisted
through reboots.
Also, if the installation of the carrier vvm causes the dialer vvm to
become disabled, do not send the deactivate sms message, as the carrier
vvm app may need it.

Bug: 21126480
Change-Id: I9ffa53c17663f5f7b324ca44a49b695ffd3646a8