Migrate 'csp_enabled' to CarrierConfigManager.
This variable controls whether SIM cards are allowed to disable manual
carrier selection, so it makes more sense for it to be managed as
carrier config.

Currently the variable is only used in *DEVICE* overlays for specific
networks, but we are dropping the device-overlay so that this simply
applies to those same networks.

Change-Id: I9ab91f79b55a88e725d789a942265da1afc3b979