Reset user changes to voicemails if the sync doesn't go through.
If a delete/read command doesn't get properly synced to the server,
simply set it back to the previous state so the user can try to
read/delete it later.
Also run a sync if signal goes back to "in service" since something may
have been missed while the signal was down.

Bug: 21978479
Change-Id: I1e40a7ff1d64f981be65516beff8f95b06fc1647