Do not retry if port number is unavailable.
If the port number is not available/can't be parsed, it means that the
IMAP credentials were incorrectly parsed or are unavailable. If this
happens, we should not retry IMAP requests.
Also retry sending activate sms if the source is not saved to the
OmtpVvmSourceManager instead of checking if the source is registered in
the voicemail_status table because the OmtpVvmSourceManager is the most
reliable place to check if the status sms has been received.

Bug: 22209621
Change-Id: I33d2bdccfb42454aebb0d1927879b4ebf23a5b7b