Reissue MWI dismissal when restoring cellular signal.
A race condition can happen if a voicemail message is received when the
signal is down. When the signal comes up again, the phone state listener
for visual voicemail will update the voicemail status table to indicate
that visual voicemail is available, at the same time, the MWI updates
with the new voicemail. Issue an update to the MWI after the voicemail
status table is updated.
- Move VvmPhoneStateListener to com.android.phone because
  NotificationMgr is only visible to the package

Bug: 22375837
Change-Id: I89b10be981da958427ec708bac3962aae04f2018