Always create CdmaPhoneCallState object in PhoneGlobals.
The CdmaPhoneCallState object used to be created on detection
of a new RAT but that code as removed in ag/703176.  So now
CDMA provisioning would crash as a null object would be dereferenced.
The lowest risk fix is to create the CdmaPhoneCallState object
on PhoneGlobals creation to guarantee that it is always there.
The worst case is that it goes unused.

Bug: 22634609
Change-Id: Ia686584476db9ce3f07fc54f3e849ad22ff7654b