Set CALL_SUBJECT capability on phone accts if instant lettering available.
- Modify TelecomAccountRegistry to check for carrier config to see if
instant lettering is available and set this on the phone accounts if
it is.
- Cleaned up some un-necessary code in isCarrierVideoPauseSupported (the
old code from prior to using carrier config manager was still present and
was not used).
- Added a Map to TelephonyConnection which provides a mapping between keys
in ImsConfig (or other Telephony sources) to their equivalent in
Telecom.  Why?  The extra for instant lettering is "DisplayText" in
ImsConfig, for example.  The RIL uses this value internally so we can't
change it.  This code remaps that extra to
android.telecom.EXTRA_CALL_SUBJECT, a properly qualified extra name which
can be exposed to incall.
the RIL

Bug: 22685114
Change-Id: I4e4e9931fa164ef576e99ef059f0c0bf2e15c909