Fix issue where conference timer resets to 0 when CEP enabled.
The problem was due to the fact that when we have children for a conference,
Telecom uses the oldest child's start time.  In the case of the CEP
participants all had no defined start time so they defaulted to using
the time when the conference was started.

Changed so that the child connections get the ImsConference's start time
(which is the start time of the oldest original call merged to form
the conference).

Bug: 23381612
Change-Id: I95766c113e7cfe2fa3b80a616bdfd9417b744dc3