Support filtering out conference originator from manage conference list.
Some carriers will include the conference originator in the conference
event package data.  This means that my own phone number can show up in
the "manage conference" screen.

Added a method in TelecomAccountRegistry to get the device phone number
given a phone account.

In ImsConference, when adding new participants add a check to see if the
participant # is the same as the conference host's number.

Also, caching and reusing the phone account handle instead of re-creating
it each time.

Bug: 22629231
Change-Id: I76067f330482c1fa2562bd89dfad659850da428f