Reformat conference event package addresses as a TEL: URI, if possible.
Conference event package data contains the phone numbers of all
participants in a conference, but is formatted as a SIP Uri.  This is
problematic as Telecom will not be able to look up contact info for the
URI, and InCall will not format the phone number appropriately.

Sip URIs can represent the phone number in a few ways, so we attempt to
extract the phone number and to reformat as a TEL uri.

Bug: 22629231
Change-Id: I3f9f3934d5db1c07547127fb9e6be4c415c7e056