Fix concurrency issue when processing conference event packages.
In ImsConference, the "handleConferenceParticipantsUpdate" method is
called in response to the "onConferenceParticipantsChanged" callback from
 the IMS stack.  Although ImsConference stores the participants in a
thread-safe collection, the code which accesses this collection when
performing the update could result in duplicates in multithreading
scenarios.

The change looks big about, but really it's just wrapping the body of
"handleConferenceParticipantsUpdate" in a synchronized block.

Bug: 23482867
Change-Id: I44317277f3cccff786f1c696b4f441a52dbd1b89