Double-check SIP state before issuing a swap-calls command.
Our code relied on the state of SipConnection to prevent us from holding
a HELD call and unholding an ACTIVE call.  However, while the action is
in progress, the SipConnection state can be slightly delayed so we
double-check the internal state of the lower connection prior to issuing
a switchHoldingAndActive command.

Bug: 24007856
Change-Id: I617ef215e66f6b9324dc17cb1b9832cc35390f17