Disallow call unhold if the fg call is dialing.
When a SIP call is in dialing state, it cannot be held. Hence ignore
switching of fg/bg calls when there are multiple calls and the
foreground call is still in dialing state. This would prevent the user
from unholding the background call and hitting a crash when we try to
switch out the foreground dialing call.

BUG: 17448699
Change-Id: I816d3c3f0253811f1ae4e7aeaa138a830d0c4d25