com.android.Phone crash popup during Phone app crash
When phone app crashed during Emergency call, phone went to
ECBM but the phone object is null due to phone app crash.
Because the phone object is null, ECBM is stopped and the
stopself api called to stop the service which in turn calling
the OnDestroy resulting in below exception due to receiver not
being registered in onCreate.

Change-Id: Ibd755728021f49d2c494915e59bddfbf87166e23
Fix: 27105697