Prevent video calling when tty mode is enabled.
This code was originally in Telecom, and it really didn't belong there.
If TTY is enabled and the user attempts to make a non-emergency video
call (same check as was in Telecom), we will fail the call with a new
disconnect cause.

See screenshot in the bug if you're curious what this will look like.
This code uses the existing code in InCall to inform the user of a failed
call reason.

Bug: 20090652
Change-Id: Iadb537e63f9eaa7b2502639d58ce2abf17511f6f