Add manual sync interval.
Visual voicemail syncing is limited to once per 60 seconds, to prevent
unnecessary syncs when mutilple events fired together. The limit is
regardless of the sync completed or failed.

If the user believes extra syncs is necessary they should be able to 
bypass this restriction. 60 seconds is too long of a wait if they have
resolved the network issue in between, or just anxiously expecting a 
call. It should still be limited to once per 3 seconds to prevent 
excessive server load.

If the sync is canceled because the interval is too short, a write to
visual voicemail source will still occur, so the intent sender can
observe the source to be informed that the sync is completed.

A better approach is to have the sender send the intent with
sendOrderedBroadcast() so the intent will be processed one by one, and
the sender can register a callback at the end to handle the sync
completed event. This will not work now because the intent is handled
asynchronously and will return before the work is done. Future commit
should address this.

Bug:26799077
Change-Id: Icfb608a868fce84d038a2597d0dc707329b35145