Fix merge after a conference call goes through SRVCC.
When conference call goes through SRVCC, a new TelephonyConnection
is created. This connection does not have the field for
conference support set to true (by default the field is not set).
As a result the conference call cannot be merged to another call
after SRVCC. The fix is to set the field to true.

Change-Id: I3e27a4e7c154c70086bfc61735fd0162292ffa1e