Prevent adding a new CDMA call if know it will fail.
See code change for additional comments.

The same CDMA command is used for merging two calls, swapping between
two calls, and adding a new call on CDMA devices. The actual function of
the command depends on the context. Instead of issuing an 'add-call'
when we know the function will do something different given the
context, send back a more meaningful error.

Bug: 22850789
Change-Id: I45b51b15f3ef2f144411db5df86536704bd1201d