Disable visual voicemail prefetch on roaming
New voicemails will not be automatically downloaded if the selected
network is on roaming to avoid excessive data charge.

It will still download regardless of roaming upon receiving
FETCH_VOICEMAIL intent. This is controlled by FetchVoicemailReceiver.
For example if the uses tapped the play button in dialer then
FETCH_VOICEMAIL will be sent and the audio will still be downloaded
because it is the result of an explicit user action.

Change-Id: I4d212b04250e74423ffcdaf17454863d7716f8d1
Fixes:26833407