Use the full SIM UICC to prevent possible collisions
Currently, Telephony registers the SIM PhoneAccounts using the partial
UICC, which has the possibility of colliding with another sim UICC on
a multi-SIM device. To Telecom, it looks like there is only one
PhoneAccount. We now register the full UICC with Telecom so that
this collision does not occur.

Change-Id: I517d05c69f1f2005b917f1e3463a14a235b49ad8
Fix: 25410213