Fix emergency call logic for DSDS in SIM2 slot only
Currently, if a DSDS phone is in Airplane mode and there is a SIM only
in the second slot, Telephony will default to using the first SIM slot
when it restarts the radio to make an emergency call. Since there is an
active SIM in slot 2, the radio will not place the call on slot 1 and
fail. This change corrects the emergency call logic to use the first SIM
that has an ICC ID.

Change-Id: Ia80f5b9581aa329b39d23b0f639a96fee0b0fa14
Fix: 27797776