Add a check to see if a call has been disconnected
Previously, we would retrieve the currently active Foreground call to
determine whether or not it is a an emergency call. The utility for
finding the currently active call includes calls in the DISCONNECTED
state, however, which would keep Airplane mode from turning on even if
the call was ended.

This change adds an additional check to be sure that the call is not in the
DISCONNECTED state before determining whether or not the call was an
emergency call.

Change-Id: Id694d8b8ef66845c834e37dba5f0077c7f145b67
Fix: 27942284