Iterate through Phones when routing Emergency Call
Iterate through the Phones directly instead of iterating based on the
Sub ID. If there are no SIMs in the device, then this approach did not
work historically (and would always use the first Phone).

Also, re-order the logic for getFirstPhoneForEmergencyCall to first
consider the user's voice calling preference, instead of always using
Slot 0.

Change-Id: Ib77347373ed98b138effe3c4735edb4eaa650aa5
Fixes: 28598339