Changes the order of operations for E911 Calls
Instead of first determining which Phone the call will be placed on and
then turning on that particular radio (when in Airplane Mode), we will
first turn off airplane mode and then wait for the radio to move to the
IN SERVICE or EMERGENCY CALLS ONLY state before deciding which Phone to
use to dial.

For calls that are not being placed in APM (including E911
Calls), the outgoing call flow will not change. The only difference will
occur when an E911 call is placed in APM. In this case, we will always
assume the call is placed as a GSM call and then switch it later if the
Call technology does indeed change.

Bug: 28200728
Change-Id: I93ad44f135777448b138656b9b70e5275719283f