Cache old VM ringtone name to avoid stutter in settings dialog
VoicemailRingtonePreference is recreated after the user selects an
option, but with an empty summary. The later async update to set the
summary will adjust the height of the preference, which makes the dialog
appear to flash. This change caches the old name in
VoicemailSettingsActivity and passes it to VoicemailRingtonePreference
upon initialization synchronously to avoid the flash

Bug: 27535183
Change-Id: Ic35fdbc6a31e7e1287f1de3ab7aa0c03222f3572