Fix unknown call logic for multi-endpoint.
When adding a new unknownn call, ensure we don't try to reuse an existing
actual call to represent an extral call.  This will happen when a call is
pulled away from the device.  We want the actual call to disconnect so it
can be logged, and a new unknown call to be created to represent the now
external version of that call.

Bug: 27458894
Change-Id: Ib5e950bce66011afab32257ff61d36b6b0d252a9