Support new user setup for VVM3
A new VVM3 SIM starts with st=N (New user). In this state some setup
is required to configure the TUI (Telephone user interface, the
traditional voicemail that the user can still call in to.)

The TUI language should be set. Spanish is selected only if the system
language is Spanish. Later CL will change this to use LocaleList.

The PIN is set to a 6 digit random number. Later CL will add an error
to prompt the user to replace the PIN. When the PIN is the generated
PIN, the "old pin" field in the change PIN dialog will be hidden
and filled with the generated PIN. If the user changed the PIN or the
server rejected the generated PIN it will be wiped from the system.

Fixes:28697797
Change-Id: Icd3f563e5ea82d1c648cf0bce19ceee56fdc4a0c