Fixed invalid MCC MNC for VVM config upon connection.
TelephonyManager.getNetworkOperator() is unreliable on CDMA network.
When the network has just connected, it could report invalid value such
as "00000", and cause an invalid config for visual voicemail to be
loaded.

TelephonyManager.getSimOperator() is more reliable. It should return a
correct value once the SIM is ready, before the network is connected.

The full cause of b/29254367 should be:
1. Full reboot:
The system is busy, activation of VVM is delayed so much that when it
is finally processing, getNetworkOperator() was stabilized and VVM config
is valid. The enabled status is written to the preferences before
ag/1126917.

2. SIM insertion:
Activation run much faster. getNetworkOperator() is called when the
network has just connected and returned a invalid value. VVM config is
deemed invalid, and the enabled status is not set which defaulted to
false before ag/1126917.

Bug: 29254367
Change-Id: I5d6dd17e7de6d04cad6f46dcddfb68d5b00df7aa