Support visual voicemail legacy mode
Legacy mode is a mode that on the carrier side visual voicemail is still
activated, but on the client side all network operations are disabled.
SMSs are still monitored so a new message SYNC SMS will be translated
to show a message waiting indicator, like traditional voicemails.

This is for carriers that does not support VVM deactivation so voicemail
can continue to function without the data cost.

Bug: 29399579
Change-Id: I0b72d6e4740aca083401ef1ae3d33976aeaa1253